Adding Algorithm for Sudoku Problem in C++
The Approach is to generate all possible configurations of numbers from 1 to 9 to fill the empty cells. Every configuration is tried one by one until the correct configuration is found, i.e. for every unassigned position fill the position with a number from 1 to 9. After filling all the unassigned position check if the matrix is safe or not. If safe print else recurs for other cases.